Refactor; move variable annotations to separate module
